requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
openai>=1.0.0
anthropic>=0.7.0
zhipuai>=2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
langchain-mcp-adapters>=0.2.1